requirements.txt
streamlit==1.28.1
sympy==1.12
google-generativeai>=0.3.0
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
